requirements.txt
fastapi==0.115.12
httpx==0.28.1
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
uvicorn==0.34.3
